requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
requests
psycopg2-binary
env
pydantic
pytz
passlib
bcrypt
cryptography
python-jose
python-dotenv


